feat(many): wIP Convert decorators to the new TypeScript 5 format
Not all of them are done, this is just withStyle, testable, WindowMessageListener.
Build and dev is
working but vitest fails with errors that look like its unable to recognize the new syntax. Likely a
configuration issue, we might have to wait for a new release of Vitest/rollup to get it fixed?